requirements.txt
# Backend
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
pydantic==2.7.1

# Database
pymongo==4.7.1
motor==3.4.0
dnspython==2.6.1
certifi

# ML
tensorflow==2.16.1
ultralytics==8.2.18
numpy==1.26.4
Pillow==10.3.0
opencv-python-headless==4.9.0.80

# LLM
openai==1.30.1
groq

# PDF
reportlab==4.2.0
matplotlib==3.9.0

# Frontend
streamlit==1.35.0
requests==2.32.2
streamlit-option-menu==0.3.12
plotly==5.22.0
pandas==2.2.2